Combine BookUploadController into BookController
